requirements.txt
# Core Django dependencies
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-otp==1.3.0
qrcode==7.4.2
cryptography==41.0.8
bcrypt==4.1.2

# API & Validation
django-filter==23.5
django-ratelimit==4.1.0
requests==2.31.0
user-agents==2.2.0

# Development
django-debug-toolbar==4.2.0
black==23.12.1
flake8==7.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.39.2

# Utilities
Pillow==10.1.0
python-decouple==3.8
celery==5.3.4
django-celery-beat==2.5.0